pyproject.toml
[tool.poetry]
name = "flash"
version = "0.1.0"
description = "Flash BE Server"
authors = ["Rob Harrigan <rob.harrigan89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.9"
djangorestframework = "^3.14.0"
stripe = "^7.12.0"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-allauth = "^0.60.1"
dj-rest-auth = "^5.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-django = "^4.7.0"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"